+++ client/src/scripts/IndexedDB/Archive/index.ts
@@ -3,6 +3,7 @@ import { getAllArchive } from "./getAllArchive"
 import { getArchive } from "./getArchive"
 import { setArchive } from "./setArchive"
 import { deleteArchive } from "./deleteArchive"
+import { existsArchive } from "./existsArchive"
 import * as find from "./findArchive"
 
 export {
@@ -11,5 +12,6 @@ export {
   getArchive,
   setArchive,
   deleteArchive,
+  existsArchive,
   find
 }